added support for 'update --auto-cache'
This option automatically initializes the specified auto-cache by
cloning the repository with --bare to the specified
location into a subdirectory named according to remote url (hashed).
The auto-caching is also implemented for the clone of submodules.

After cloning a repository into the workspace (using the local cache),
the remote URL is updated to point to the actual remote. This means
that when running west update, the repository in the workspace is
fetched and updated from the original remote source.

During update first the local auto-cache is fetched. In a next step the
workspace repository fetches from the auto-cache.

Signed-off-by: Thorsten Klein <thorsten.klein@bshg.com>