auto-cache: make sure a revision is checked out after clone
When a remote default branch is renamed and the auto-cache is updated
with remote, the auto-cache becomes corrupt as HEAD points to an invalid
remote HEAD. This leads to failing `west update`. In older git version a
warning was emitted ("warning: remote HEAD refers to nonexistent ref,
unable to checkout"), but in newer git versions, the issue is fixed.